kernel-devsrc: ppc32: fix reproducibility
In a similar manner to the arm/arm32 reprodicibility fixes,
we can also fix ppc32.

The file .vdso32-offsets.h.cmd has captured paths, but we don't
need it on target or SDKs to regenerate a build enviroment.

We add it to our vdso-offsets list of files to delete and we
no longer have files patckaged with buildpaths.

(From OE-Core rev: 1b06b29533733cb42589698a066d3784af3d5f2f)

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>